requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
psycopg2  
marshmallow
pytest
langchain
ollama
